+++ src/containers/Home/index.tsx
@@ -3,35 +3,10 @@ import { dayBlockAPI } from '@/api'
 import dayjs from 'dayjs'
 import { GetDailyBlocksOnWeekResponse } from '@/api/types/base.types'
 import Tabs from '@/components/Tabs'
-import Button from '@/components/Button'
-import { AddIcon } from '@/components/Icons'
-import { BASE_URL } from '@/constants/urls'
 import ProfileHeader from './ProfileHeader'
 import CalendarPanel from './CalendarPanel'
 import DailyBlockPanel from './DailyBlockPanel'
 import BlockList from './BlockList'
-
-import rnWebViewBridge from '@/utils/react-native-webview-bridge/new-webview/rnWebViewBridge'
-
-const AddButton = ({
-  text,
-  onClick,
-}: {
-  text: string
-  onClick: () => void
-}) => {
-  return (
-    <Button
-      color="gray"
-      fontWeight="bold"
-      className="py-[11px]"
-      onClick={onClick}
-    >
-      <AddIcon className="!fill-textGray-200" width={18} height={18} />
-      {text}
-    </Button>
-  )
-}
 
 const Home = () => {
   const [data, setData] = useState<GetDailyBlocksOnWeekResponse | null>(null)
@@ -45,20 +20,6 @@ const Home = () => {
     }
     getDailyBlocks()
   }, [today])
-
-  const handleAddSavedBlock = () => {
-    rnWebViewBridge.open({
-      key: 'savedBlock',
-      url: `${BASE_URL}/blocks/saved`,
-    })
-  }
-
-  const handleAddNewBlock = () => {
-    rnWebViewBridge.open({
-      key: 'newBlock',
-      url: `${BASE_URL}/blocks/new`,
-    })
-  }
 
   if (!data) return null
   const { user, dailyBlocks } = data
@@ -81,11 +42,6 @@ const Home = () => {
       </Tabs>
 
       <BlockList />
-
-      <div className="flex gap-2">
-        <AddButton text="블럭 불러오기" onClick={handleAddSavedBlock} />
-        <AddButton text="새 블럭 만들기" onClick={handleAddNewBlock} />
-      </div>
     </div>
   )
 }
